[memory refactor][4/n] Update the client-sdk test for RAG
